init: compile out slave_core_init on UP arch
Compiles out slave_core_init on UP architecure.
